requirements.txt
opencv-python==4.9.0.80
numpy~=1.26.4
Pillow==10.0.0
pytesseract==0.3.10
pdf2image==1.16.3
reportlab~=4.0.8
openai~=2.7.2